pyproject.toml
[project]
name = "sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "groq>=1.1.2",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
]
